refactor(agenttelemetry): rename per iglendd/georgeshachem review feedback
Implements reviewer feedback from PR #50607 (excluding the
stack-frame-count reduction on errortracking_sender.go:101):

- georgeshachem: `logsPayloadType` = "agent-logs" (was "logs") so COAT
  intake can distinguish error-tracking records from tracer logs.
- iglendd: rename public Component method SubmitErrorRecord →
  SubmitErrorLog; cascade to impl, def, cmd, tests, and doc comments.
- iglendd: rename internal helpers sendLogsTypedBatch → sendLogsBatch
  and sendSerializedPayload → sendPayload; cascade to all call sites,
  comments, and tests.
- iglendd: rename enrichment helpers errorLogToLog → enrichErrorLog and
  symbolizeStack → symbolizeStackFrames; cascade to tests.
- iglendd: change Bouncer key type uintptr → uint64 throughout
  (bouncer.go, bouncer_test.go, handler.go) so hashPCs returns
  h.Sum64() directly without the 32-bit-truncating cast.

All tests green: comp/core/agenttelemetry/impl/... and
pkg/util/log/errortracking/...

+++ pkg/util/log/errortracking/handler.go
@@ -165,11 +165,11 @@ func (h *Handler) Handle(_ context.Context, r slog.Record) error {
 	return nil
 }
 
-// hashPCs returns a 64-bit FNV-1a hash of the captured stack PCs,
-// truncated to uintptr. The hash is the bouncer key — two records
-// reaching the same terminal function from different call stacks
-// produce different hashes and are NOT deduped together.
-func hashPCs(pcs []uintptr) uintptr {
+// hashPCs returns a 64-bit FNV-1a hash of the captured stack PCs.
+// The hash is the bouncer key — two records reaching the same terminal
+// function from different call stacks produce different hashes and are
+// NOT deduped together.
+func hashPCs(pcs []uintptr) uint64 {
 	h := fnv.New64a()
 	var buf [8]byte
 	for _, pc := range pcs {
@@ -178,7 +178,7 @@ func hashPCs(pcs []uintptr) uintptr {
 		}
 		h.Write(buf[:])
 	}
-	return uintptr(h.Sum64())
+	return h.Sum64()
 }
 
 // WithAttrs is a required slog.Handler interface method. Attrs are not
